+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/core/CardCrawlGame/LoadPlayerSaves.java
@@ -1,17 +1,17 @@
 package basemod.patches.com.megacrit.cardcrawl.core.CardCrawlGame;
 
-import java.util.Map;
-
+import basemod.BaseMod;
+import basemod.abstracts.CustomSavableRaw;
+import basemod.patches.com.megacrit.cardcrawl.saveAndContinue.SaveFile.ModSaves;
 import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.potions.AbstractPotion;
 import com.megacrit.cardcrawl.relics.AbstractRelic;
 
-import basemod.BaseMod;
-import basemod.abstracts.CustomSavableRaw;
-import basemod.patches.com.megacrit.cardcrawl.saveAndContinue.SaveFile.ModSaves;
+import java.util.Map;
 
 @SpirePatch(clz=CardCrawlGame.class, method="loadPlayerSave")
 public class LoadPlayerSaves
@@ -34,6 +34,16 @@ public class LoadPlayerSaves
         for (AbstractRelic relic : AbstractDungeon.player.relics) {
             if (relic instanceof CustomSavableRaw) {
                 ((CustomSavableRaw)relic).onLoadRaw(modRelicSaves == null || i >= modRelicSaves.size() ? null : modRelicSaves.get(i));
+            }
+            i++;
+        }
+
+        // Potions
+        ModSaves.ArrayListOfJsonElement modPotionSaves = ModSaves.modPotionSaves.get(CardCrawlGame.saveFile);
+        i = 0;
+        for (AbstractPotion potion : AbstractDungeon.player.potions) {
+            if (potion instanceof CustomSavableRaw) {
+                ((CustomSavableRaw)potion).onLoadRaw(modPotionSaves == null || i >= modPotionSaves.size() ? null : modPotionSaves.get(i));
             }
             i++;
         }