+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/saveAndContinue/SaveFile/ModSaves.java
@@ -23,4 +23,6 @@ public class ModSaves
     public static SpireField<ArrayListOfJsonElement> modCardSaves = new SpireField<>(() -> null);
     @SerializedName("basemod:mod_relic_saves")
     public static SpireField<ArrayListOfJsonElement> modRelicSaves = new SpireField<>(() -> null);
+    @SerializedName("basemod:mod_potion_saves")
+    public static SpireField<ArrayListOfJsonElement> modPotionSaves = new SpireField<>(() -> null);
 }